added authentication

+++ internal/products/handler.go
@@ -2,6 +2,8 @@ package products
 
 import (
 	"github.com/Picus-Security-Golang-Bootcamp/bitirme-projesi-ybaspinar/internal/models"
+	jwtHelper "github.com/Picus-Security-Golang-Bootcamp/bitirme-projesi-ybaspinar/pkg/JWT"
+	"github.com/Picus-Security-Golang-Bootcamp/bitirme-projesi-ybaspinar/pkg/config"
 	"github.com/Picus-Security-Golang-Bootcamp/bitirme-projesi-ybaspinar/pkg/pagination"
 	_ "github.com/Picus-Security-Golang-Bootcamp/bitirme-projesi-ybaspinar/pkg/pagination"
 	"github.com/gin-gonic/gin"
@@ -10,35 +12,43 @@ import (
 
 type productHandler struct {
 	repo *ProductRepo
+	cfg  *config.Config
 }
 
-//TODO: add authentication
 func (h *productHandler) create(context *gin.Context) {
 	var product models.Product
-	if err := context.ShouldBindJSON(&product); err != nil {
-		context.JSON(400, gin.H{"error": err.Error()})
-		return
+	token := context.GetHeader("Authorization")
+	decodedClaims := jwtHelper.VerifyToken(token, h.cfg.JWTConfig.SecretKey)
+	if decodedClaims.IsAdmin {
+		if err := context.ShouldBindJSON(&product); err != nil {
+			context.JSON(400, gin.H{"error": err.Error()})
+			return
+		}
+		if err := h.repo.Create(&product); err != nil {
+			context.JSON(500, gin.H{"error": err.Error()})
+			return
+		}
+		context.JSON(http.StatusOK, product)
+	} else {
+		context.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 	}
-	if err := h.repo.Create(&product); err != nil {
-		context.JSON(500, gin.H{"error": err.Error()})
-		return
-	}
-	context.JSON(http.StatusOK, product)
 }
 
-//TODO: add authentication
 func (h *productHandler) delete(context *gin.Context) {
 	id := context.Param("id")
-	println("printing id: ", id)
-	println(id)
-	if err := h.repo.Delete(id); err != nil {
-		context.JSON(500, gin.H{"error": err.Error()})
-		return
+	token := context.GetHeader("Authorization")
+	decodedClaims := jwtHelper.VerifyToken(token, h.cfg.JWTConfig.SecretKey)
+	if decodedClaims.IsAdmin {
+		if err := h.repo.Delete(id); err != nil {
+			context.JSON(500, gin.H{"error": err.Error()})
+			return
+		}
+		context.JSON(http.StatusOK, gin.H{"message": "Product deleted"})
+	} else {
+		context.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 	}
-	context.JSON(http.StatusOK, gin.H{"message": "Product deleted"})
 }
 
-//TODO: add authentication
 func (h *productHandler) search(context *gin.Context) {
 	pageIndex, pageSize := pagination.GetPaginationParametersFromRequest(context)
 	products, totalCount := h.repo.FuzzySearchSkuAndNameAndId(context.Param("q"), pageIndex, pageSize)
@@ -47,18 +57,25 @@ func (h *productHandler) search(context *gin.Context) {
 	context.JSON(http.StatusOK, products)
 }
 
-//TODO: add authentication
 func (h *productHandler) update(context *gin.Context) {
 	var product models.Product
-	if err := context.ShouldBindJSON(&product); err != nil {
-		context.JSON(400, gin.H{"error": err.Error()})
-		return
+	token := context.GetHeader("Authorization")
+	decodedClaims := jwtHelper.VerifyToken(token, h.cfg.JWTConfig.SecretKey)
+	if decodedClaims.IsAdmin {
+		if err := context.ShouldBindJSON(&product); err != nil {
+			context.JSON(400, gin.H{"error": err.Error()})
+			return
+		}
+
+		if err := h.repo.Update(&product); err != nil {
+			context.JSON(500, gin.H{"error": err.Error()})
+			return
+		}
+		context.JSON(http.StatusOK, product)
+	} else {
+		context.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 	}
-	if err := h.repo.Update(&product); err != nil {
-		context.JSON(500, gin.H{"error": err.Error()})
-		return
-	}
-	context.JSON(http.StatusOK, product)
+
 }
 
 func (h *productHandler) getAll(context *gin.Context) {
@@ -69,8 +86,8 @@ func (h *productHandler) getAll(context *gin.Context) {
 	context.JSON(http.StatusOK, paginatedResponse)
 }
 
-func NewProductHandler(r *gin.RouterGroup, repo *ProductRepo) {
-	h := &productHandler{repo: repo}
+func NewProductHandler(r *gin.RouterGroup, repo *ProductRepo, cfg *config.Config) {
+	h := &productHandler{repo: repo, cfg: cfg}
 
 	r.POST("/create", h.create)
 	r.GET("/", h.getAll)
